refactor(@konfik/cli): move command-related code into separate directory
